+++ src/main/java/com/selfrunner/gwalit/domain/board/dto/response/ReplyRes.java
@@ -3,7 +3,7 @@ package com.selfrunner.gwalit.domain.board.dto.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.selfrunner.gwalit.domain.board.entity.Reply;
 import com.selfrunner.gwalit.domain.member.entity.Member;
-import com.selfrunner.gwalit.domain.member.entity.MemberType;
+import com.selfrunner.gwalit.domain.member.enumerate.MemberType;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
