+++ src/main/java/com/selfrunner/gwalit/domain/lecture/dto/request/PostStudentReq.java
@@ -1,8 +1,8 @@
 package com.selfrunner.gwalit.domain.lecture.dto.request;
 
 import com.selfrunner.gwalit.domain.member.entity.Member;
-import com.selfrunner.gwalit.domain.member.entity.MemberState;
-import com.selfrunner.gwalit.domain.member.entity.MemberType;
+import com.selfrunner.gwalit.domain.member.enumerate.MemberState;
+import com.selfrunner.gwalit.domain.member.enumerate.MemberType;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
