+++ src/main/java/com/selfrunner/gwalit/domain/member/entity/MemberAndLecture.java
@@ -3,6 +3,7 @@ package com.selfrunner.gwalit.domain.member.entity;
 import com.selfrunner.gwalit.domain.lecture.dto.request.PatchColorReq;
 import com.selfrunner.gwalit.domain.lecture.dto.request.PatchNameReq;
 import com.selfrunner.gwalit.domain.lecture.entity.Lecture;
+import com.selfrunner.gwalit.domain.member.enumerate.MemberType;
 import com.selfrunner.gwalit.global.common.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
