--- src/main/java/com/selfrunner/gwalit/domain/member/entity/MemberType.java
@@ -1,5 +1,0 @@
-package com.selfrunner.gwalit.domain.member.entity;
-
-public enum MemberType {
-    TEACHER, STUDENT
-}
