+++ src/main/java/com/selfrunner/gwalit/domain/member/enumerate/MemberType.java
@@ -1,0 +1,5 @@
+package com.selfrunner.gwalit.domain.member.enumerate;
+
+public enum MemberType {
+    TEACHER, STUDENT
+}
