+++ src/main/java/com/selfrunner/gwalit/domain/member/repository/MemberRepository.java
@@ -1,7 +1,6 @@
 package com.selfrunner.gwalit.domain.member.repository;
 
 import com.selfrunner.gwalit.domain.member.entity.Member;
-import com.selfrunner.gwalit.domain.member.entity.MemberType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
